Bug 1314614 - Replace GetCurrentProcessId calls by getpid. r=h4writer
UltraBlame original commit: 72eddba8641ac119f72b4d340591b312010e278a

+++ js/src/vm/CodeCoverage.h
@@ -162,7 +162,7 @@ class LCovRuntime
 
     
     
-    size_t pid_;
+    uint32_t pid_;
